split up execution context utils
